Master Theorem
The Master Theorem proved to be a very important theorem in computer science and theoretical mathematics. It helps one analyze algorithms and functions that rely on recursion/iterations. The Master Theorem would be useful to programmers since the Master Theorem helps portray the runtimes of algorithms. The runtime of an algorithm is important to a programmer, since the best, most efficient algorithm is always the primary goal when implementing one.